requirements.txt
aiobotocore==2.5.1
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.164
botocore==1.29.161
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
fsspec==2023.6.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
numpy==1.25.0
pandas==2.0.3
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3fs==2023.6.0
s3transfer==0.6.1
six==1.16.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==1.26.16
wrapt==1.15.0
yarl==1.9.2
